+++ src/entities/offer-review/model/constants.ts
@@ -1,0 +1,4 @@
+export const MONTHS = [
+  'January', 'February', 'March', 'April', 'May', 'June',
+  'July', 'August', 'September', 'October', 'November', 'December'
+];
