+++ src/shared/constants/generalConstants.ts
@@ -1,0 +1,1 @@
+export const NUMBER_OF_OFFERS = 322;
