+++ src/shared/constants/index.ts
@@ -1,1 +1,2 @@
-export * from './constants';
+export * from './generalConstants';
+export * from './routingConstants';
