+++ src/shared/constants/routingConstants.ts
@@ -4,5 +4,3 @@ export enum AppRoute {
   Favorites = '/favorites',
   Offer = '/offer/:id',
 }
-
-export const NUMBER_OF_OFFERS = 322;
