+++ src/shared/helpers/offerHelpers.ts
@@ -1,0 +1,7 @@
+export const getRatingPercent = (ratingNumber: number) => {
+  const ratingPercentValue = ratingNumber / 5 * 100;
+
+  return `${ratingPercentValue}%`;
+};
+
+
