+++ src/shared/index.ts
@@ -1,4 +1,5 @@
 export * from './components';
 export * from './constants';
+export * from './helpers';
 export * from './mocks';
 export * from './types';
