+++ src/shared/mocks/offerMocks.ts
