+++ src/shared/types/index.ts
@@ -1,2 +1,2 @@
-export * from './types';
+export * from './authTypes';
 export * from './offerTypes';
